refactor(auth): extract hook callbacks to src/lib/auth/hooks.ts (CP-4)
Extracts 11 long callback bodies (sendResetPassword, onPasswordReset,
beforeDelete validation, databaseHooks for user/session create, organization
plugin lifecycle hooks) as named functions. Keeps auth.api.* calls inline in
lib/auth.ts to avoid circular imports.
